// src/types/strapi.ts
export interface Customization {
  palette: { [colorName: string]: { main: string } };
  logoName?: string;
  footerBannerUrl?: string;
  typography?: string;
}

type WidgetConfigProps = Omit<WidgetConfig, 'integrator'>;

export interface PartnerTheme {
  config: WidgetConfigProps;
  customization?: Customization;
}